+++ src/main/java/com/uwetrottmann/tmdb/services/TvService.java
@@ -1,5 +1,6 @@
 package com.uwetrottmann.tmdb.services;
 
+import com.uwetrottmann.tmdb.entities.AppendToResponse;
 import com.uwetrottmann.tmdb.entities.Credits;
 import com.uwetrottmann.tmdb.entities.ExternalIds;
 import com.uwetrottmann.tmdb.entities.Images;
@@ -20,11 +21,13 @@ public interface TvService {
      *
      * @param tmdbId A themoviedb id.
      * @param language <em>Optional.</em> ISO 639-1 code.
+     * @param appendToResponse <em>Optional.</em> extra requests to append to the result.
      */
     @GET("/tv/{id}")
     TvShowComplete tv(
             @Path("id") int tmdbId,
-            @Query("language") String language
+            @Query("language") String language,
+            @Query("append_to_response") AppendToResponse appendToResponse
     );
     
     /**